requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.11
firecrawl==1.15.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
joblib==1.4.2
narwhals==1.30.0
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.68.0
packaging==24.2
pandas==2.2.3
plotly==6.0.0
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.12.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
sse-starlette==2.2.1
starlette==0.46.1
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
websockets==15.0.1
